Split up library a bit more
I found that require.main is not consistent for a number of reasons, and
can't be trusted. This will need to be replicated on the EA client and
other libraries.

Basically, it outputs one script that will be used in the browser, and
one script that will be used normally if you are installing and
including it in as a module.

This is the most explicit way of telling the difference for now, until
there is a similar `require.main` for ES modules.